Add automatic build wrapper with name scope

+++ keras_core/backend/tensorflow/image.py
@@ -47,7 +47,7 @@ def resize(
             resized = tf.transpose(resized, (0, 3, 1, 2))
         elif len(image.shape) == 3:
             resized = tf.transpose(resized, (2, 0, 1))
-    return tf.cast(resized, image.dtype)
+    return resized
 
 
 AFFINE_TRANSFORM_INTERPOLATIONS = (
